+++ priv/repo/migrations/20230920043856_add_item_details_table.exs
@@ -5,7 +5,7 @@ defmodule JustStore.Repo.Migrations.AddItemDetailsTable do
     create table(:item_details) do
       add :name, :string
       add :description, :string
-      add :price, :decimal
+      add :price, :float
       timestamps()
     end
   end
